+++ index.html
@@ -8,7 +8,7 @@
   <title>Кексобукинг</title>
 </head>
 <body>
-  <script defer src="/js/main.js"></script>
+  <script defer src="/js/main.js" type="module"></script>
   <main>
     <div class="promo">
       <h1 class="promo__title visually-hidden">Keksobooking. Кексы по соседству</h1>
